refactor: add utils folder
moved scanner and ocr into utils folder
