+++ app/src/main/java/com/github/onlynotesswent/ui/overview/AddNote.kt
@@ -39,7 +39,7 @@ import androidx.compose.ui.unit.dp
 import com.github.onlynotesswent.R
 import com.github.onlynotesswent.model.note.Note
 import com.github.onlynotesswent.model.note.NoteViewModel
-import com.github.onlynotesswent.model.scanner.Scanner
+import com.github.onlynotesswent.utils.Scanner
 import com.github.onlynotesswent.model.users.UserViewModel
 import com.github.onlynotesswent.ui.navigation.NavigationActions
 import com.google.firebase.Timestamp
@@ -47,10 +47,10 @@ import com.google.firebase.Timestamp
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun AddNoteScreen(
-    navigationActions: NavigationActions,
-    scanner: Scanner,
-    noteViewModel: NoteViewModel,
-    userViewModel: UserViewModel
+  navigationActions: NavigationActions,
+  scanner: Scanner,
+  noteViewModel: NoteViewModel,
+  userViewModel: UserViewModel
 ) {
 
   var title by remember { mutableStateOf("") }
