+++ app/src/test/java/com/github/onlynotesswent/model/scanner/ScannerTest.kt
@@ -17,6 +17,7 @@ import androidx.activity.result.IntentSenderRequest
 import androidx.activity.result.contract.ActivityResultContract
 import androidx.core.content.FileProvider
 import androidx.core.content.FileProvider.getUriForFile
+import com.github.onlynotesswent.utils.Scanner
 import com.google.android.gms.tasks.OnFailureListener
 import com.google.android.gms.tasks.OnSuccessListener
 import com.google.android.gms.tasks.Task
